Moved Risk struct to models
